Target as property
I realized it makes more sense for the target to be a property of each
binary. Dependencies are specified using a list of properties. There it
makes sense to "apply this target to all of these dependencies below". This
creates a slightly different syntax, but it makes sense that it differs for
the two different cases.